+++ packages/uikit-react-native-foundation/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.0](https://github.com/sendbird/sendbird-uikit-react-native/compare/v3.6.0...v3.7.0) (2024-08-08)
+
+### Features
+
+- updated sample React version to 0.74.3 ([0e32587](https://github.com/sendbird/sendbird-uikit-react-native/commit/0e32587b51b07b160c72d393e044fb6532867d1f))
+
+## [3.6.0](https://github.com/sendbird/sendbird-uikit-react-native/compare/v3.5.4...v3.6.0) (2024-07-09)
+
+### Features
+
+- added thread type in GroupChannel ([5e815ea](https://github.com/sendbird/sendbird-uikit-react-native/commit/5e815ea73efec563c50844ece052fae0fb0694b7))
+
 ## [3.5.4](https://github.com/sendbird/sendbird-uikit-react-native/compare/v3.5.3...v3.5.4) (2024-06-13)
 
 **Note:** Version bump only for package @sendbird/uikit-react-native-foundation